1.Event修复 2.ObjectSystem跟Event不再需要加标记，因为基类已经加了 3.拆分显示层跟逻辑层，显示层的组件放到View里面 4.publish改成支持异步的

+++ Server/App/Program.cs
@@ -29,13 +29,12 @@ namespace ET
 				
 				Game.Scene.AddComponent(options);
 				
-				IdGenerater.Process = (byte) options.Process;
+				IdGenerater.Process = options.Process;
 				
 				LogManager.Configuration.Variables["appIdFormat"] = $"{Game.Scene.Id:0000}";
 				
 				Log.Info($"server start........................ {Game.Scene.Id}");
 
-				// 先加这里，后面删掉
 				Game.EventSystem.Publish(new EventType.AppStart());
 				
 				while (true)
